pyproject.toml
[project]
name = "deep-eval-poc"
version = "0.1.0"
description = "A proof-of-concept repository for testing DeepEval framework"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "deepeval (>=0.21.0)",
    "openai (>=1.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"

